+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Added
+
+- Completion to close the current tag when typing `</` (like `</tag>`).
+- Snippets for new documents.
+
+### Fixed
+
+- Broken snippets
+
 ## [0.16.0] - 2024-02-10
 
 ### Added